requirements.txt
PyQt6>=6.4
PyYAML==6.0.1
python-barcode==0.15.1
Pillow==10.4.0
qrcode==7.4.2
